Simplified the Env class that links to Env enum
Since the Env class in portal/environment is a class representation of
the enum Env in the apollo core project I linked them better together.
This means if important changes happen in the enum it is automatically
reflected to the class representation.

Additionall I simplified the `transformEnv` method inside the Env class.
The switch will be never reached since all those values are already
matched in the `Env.valueOf(envName)`.

Further more I added more javadoc and tests to boost up the coverage