`ConfigFileFormat#Properties` are now fully compatible with themselves
Previously `ConfigFileFormat#Properties` were not fully compatible by
themself which meant the function `ConfigFileFormat.isPropertiesCompatible`
was not really meaning full. PropertiesConfigFiles are now designed to be
compatible bythemselves by implementing `PropertiesCompatibleConfigFile`.
This might be usefull in further development.

Currently `DefaultConfigFactory.create` will not create a properties
compatible repository since this is not really usefull and a
localRepositry might be easier to use in further development.